Fixed deletion and improved infa fks tracking.

+++ Pdc.Infrastructure/Repositories/CompetencyRepository.cs
@@ -4,11 +4,9 @@ using Microsoft.EntityFrameworkCore;
 using Pdc.Domain.Interfaces.Repositories;
 using Pdc.Domain.Models.CourseFramework;
 using Pdc.Domain.Models.MinisterialSpecification;
-using Pdc.Domain.Models.Security;
 using Pdc.Domain.Models.Versioning;
 using Pdc.Infrastructure.Data;
 using Pdc.Infrastructure.Entities.CourseFramework;
-using Pdc.Infrastructure.Entities.Identity;
 using Pdc.Infrastructure.Entities.MinisterialSpecification;
 using Pdc.Infrastructure.Entities.Versioning;
 using Pdc.Infrastructure.Exceptions;
@@ -32,40 +30,46 @@ public class CompetencyRepository : ICompetencyRepository
         return _mapper.Map<List<MinisterialCompetency>>(entities);
     }
 
-    public async Task<MinisterialCompetency> Add(ProgramOfStudy program, MinisterialCompetency competency, User currentUser)
+    public async Task<MinisterialCompetency> Add(ProgramOfStudy program, MinisterialCompetency competency)
     {
+        ProgramOfStudyEntity programEntity = await FindProgramOfStudy(program.Code);
         CompetencyEntity entity = await _context.Competencies
             .Persist(_mapper)
             .InsertOrUpdateAsync(competency);
-        // TODO move that into the usecase logic.
-        IdentityUserEntity? user = await _context.Users.FirstOrDefaultAsync(x => x.Id == currentUser.Id);
-        if (user is null)
-        {
-            throw new EntityNotFoundException(nameof(IdentityUserEntity), currentUser?.Id != null ? currentUser.Id : "id not found");
-        }
-        entity.SetCreatedBy(user);
-        // End of TODO.
-        var addedEntity = _context.Competencies.Add(entity);
-        ProgramOfStudyEntity programEntity = await FindProgramOfStudy(program.Code);
-        programEntity.Competencies.Add(addedEntity.Entity);
+        entity.ProgramOfStudy = programEntity;
         await _context.SaveChangesAsync();
         return _mapper.Map<MinisterialCompetency>(entity);
     }
 
     public async Task<MinisterialCompetency> Update(MinisterialCompetency competency)
     {
-        CompetencyEntity existingCompetency = await FindEntityByCode(competency.Code);
-        MarkAllRemovedElementsForDeletion(competency, existingCompetency);
+        await MarkAllRemovedElementsForDeletion(competency);
+
         CompetencyEntity entity = await _context.Competencies
             .Persist(_mapper)
             .InsertOrUpdateAsync(competency);
-
         await _context.SaveChangesAsync();
         return _mapper.Map<MinisterialCompetency>(entity);
     }
 
-    private void MarkAllRemovedElementsForDeletion(MinisterialCompetency competency, CompetencyEntity existingCompetency)
+    private async Task MarkAllRemovedElementsForDeletion(MinisterialCompetency competency)
     {
+        CompetencyEntity? existingCompetency = await _context.Competencies
+            .AsNoTracking()
+            .AsSplitQuery()
+            .Include(c => c.RealisationContexts)
+                .ThenInclude(rc => rc.ComplementaryInformations)
+            .Include(c => c.CompetencyElements)
+                .ThenInclude(ce => ce.ComplementaryInformations)
+            .Include(c => c.CompetencyElements)
+                .ThenInclude(ce => ce.PerformanceCriterias)
+                    .ThenInclude(pc => pc.ComplementaryInformations)
+            .Include(c => c.ProgramOfStudy)
+            .SingleOrDefaultAsync(x => x.Code == competency.Code);
+        if (existingCompetency == null)
+        {
+            throw new EntityNotFoundException(nameof(MinisterialCompetency), competency.Code);
+        }
         List<ChangeableEntity> realisationContextToDelete = FindMissingAChangeableForDeletion(competency.RealisationContexts.Cast<AChangeable>().ToList(), existingCompetency.RealisationContexts.Cast<ChangeableEntity>().ToList());
         _context.RealisationContexts.RemoveRange(realisationContextToDelete.Cast<RealisationContextEntity>().ToList());
     }
@@ -92,7 +96,7 @@ public class CompetencyRepository : ICompetencyRepository
 
     public async Task<MinisterialCompetency> FindByCode(string competencyCode)
     {
-        CompetencyEntity entity = await FindEntityByCode(competencyCode);
+        CompetencyEntity entity = await FindDetailedEntityByCode(competencyCode);
         return _mapper.Map<MinisterialCompetency>(entity);
     }
 
@@ -103,9 +107,11 @@ public class CompetencyRepository : ICompetencyRepository
             .AnyAsync();
     }
 
-    private async Task<CompetencyEntity> FindEntityByCode(string competencyCode)
+    private async Task<CompetencyEntity> FindDetailedEntityByCode(string competencyCode)
     {
-        CompetencyEntity? competency = await _context.Competencies.AsNoTracking()
+        CompetencyEntity? competency = await _context.Competencies
+            .AsSplitQuery()
+            .AsNoTracking()
             .Include(c => c.RealisationContexts)
                 .ThenInclude(rc => rc.ComplementaryInformations)
                     .ThenInclude(cr => cr.WrittenOnVersion)
@@ -135,7 +141,6 @@ public class CompetencyRepository : ICompetencyRepository
     private async Task<ProgramOfStudyEntity> FindProgramOfStudy(string code)
     {
         ProgramOfStudyEntity? program = await _context.ProgramOfStudies
-            .Include(p => p.Competencies)
             .SingleOrDefaultAsync(x => x.Code == code);
         if (program == null)
